Add waiting for interrupts in userspace.
Change the way block devices are stored in the kernel.

Add ToTrait trait that is like Any, but allows for casting into traits. The #[to_trait] macro implements this trait. Probably safe.

Add an ArcInject struct to allow attaching structures containing references to data in an Arc to the Arc itself, such as trait objects.

Run cargo fmt